==> analysis/salmon_demo/step1.create_salmon_index.ipynb <==
# Build Salmon Index with GENCODE

# In salmon documentation, it provided a tutorial for building decoy awared reference index, you should do so in real analysis. But here for the demo purpose, I will simply use the transcripts only to make the indexing process faster

See salmon tutorial here: https://combine-lab.github.io/alevin-tutorial/2019/selective-alignment/

## Download GENCODE vm25 transcript FASTA file
- See this link: https://www.gencodegenes.org/mouse/
- The tutorial is about mouse, but you can download Human files here: ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_34/

!wget ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M25/gencode.vM25.transcripts.fa.gz

## Create salmon index

### install salmon to your env if you haven't
`conda install -n genome_book -c bioconda salmon`

### you can also execute this in your command line, but here I will use the subprocess package
`salmon index -t gencode.vM25.transcripts.fa.gz -p 4 -i ./data/salmon_index --gencode`

import subprocess

p = subprocess.run('salmon index -t gencode.vM25.transcripts.fa.gz -p 4 -i ./data/salmon_index --gencode',
               shell=True, stdout=subprocess.PIPE, stdin=subprocess.PIPE, encoding='utf8')
print(p.stdout)
