==> analysis/salmon_demo/step2.PrepareMetadata.ipynb <==
# Prepare Developing Forebrain Bulk-seq (Demo)

# **All FASTQ files only contain 10000 reads form the source file downloaded below, so you can map it quickly for this demo**

## Aims of this notebook

1. prepare a clean metadata table, recording sample information for each FASTQ file
2. create soft-link to FASTQ files, the soft-link is a good way to rename those FASTQ files with a meaningful name, while keep the original file unchanged.

## Data Source
- All the data downloaded from ENCODE project [here](https://www.encodeproject.org/search/?type=Experiment&status=released&lab.title=Barbara+Wold%2C+Caltech&biosample_ontology.term_name=forebrain)
- This dataset contains polyA+ RNA-seq for developing forebrain from 8 timepoints 
    - E10.5
    - E11.5
    - E12.5
    - E13.5
    - E14.5
    - E15.5
    - E16.5
    - P0
- Each time point has two replicates, a total of 16 RNA-seq
- Some experiments are splited into multiple FASTQ files, some are not
- We start from a total of 25 fastq files from ENCODE, all experiments are single end sequencing

## Rename file and make metadata

# pandas handle table
import pandas as pd
# pathlib handle all path related
import pathlib

# Read this raw metadata table downloaded from ENCODE together with the FASTQ
metadata = pd.read_csv('../../data/DevFB/fastq/raw_metadata.tsv', sep='\t', index_col=0)
print('The dataframe shape', metadata.shape)
metadata.head()

metadata.columns

### Clean metadata, select necessary columns

# The raw metadata table contain many unnecessary columns, here we select important ones only
use_columns = [
    'Output type', 'Experiment accession', 'Biosample term id',
    'Biosample term name', 'Biological replicate(s)'
]

metadata_selected = metadata[use_columns].copy()
print('The dataframe shape', metadata_selected.shape)
metadata_selected.head()

### Concatenate timepoint information

# I manually create this sample table so we can paste the "Stage" (development time point) with the metadata table
exp_meta = pd.read_csv('../../data/DevFB/metadata/ENCODE_experiments.tsv', sep='\t', index_col=-1)
print('The dataframe shape', exp_meta.shape)
exp_meta.head()

# These Experiment accession "ENCSR..." matched with the exp_meta, 
# our goal is add the Stage column from exp_meta to the metadata_selected -- using these IDs
metadata_selected['Experiment accession'].head()

# single line of code can do this, search pandas map function 
# or see documentation here: https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.map.html
metadata_selected['Stage'] = metadata_selected['Experiment accession'].map(exp_meta['Stage'])

print('The dataframe shape', metadata_selected.shape)
metadata_selected.head()

### [OPTIONAL] Check some questions
- Examples to validate your metadata
- Demo of some complex logic check using pandas functions

print('Number of tissues', 
      metadata_selected['Biosample term name'].unique().size)
metadata_selected['Biosample term name'].unique()

print('Number of time points in each tissue')
metadata_selected.groupby('Biosample term name')['Stage'].apply(lambda i: i.unique().size)

print('Does each experiment (tissue*stage) has 2 replicates?')

# In human words, this means the number of replicates in each tissue and stage combination all equal to 2
# This is just a demo of how complex logic validation can be achieved with python.
sum(
    metadata_selected\
    .groupby(['Biosample term name', 'Stage'])\
    .apply(lambda i: i['Biological replicate(s)'].unique().size) != 2
) == 0

### Rename the files, add sample information

# create a output dir using command line
!mkdir data/fastq

# another way to create directory using pure python
output_dir = pathlib.Path('data/fastq')
output_dir.mkdir(exist_ok=True)

# rename the columns
rename_dict = {
    'Output type': 'count_type',
    'Experiment accession': 'experiment_id',
    'Biosample term id': 'bio_sample_id',
    'Biosample term name': 'tissue',
    'Biological replicate(s)': 'replicate',
    'Stage': 'dev_time'
}
metadata_selected.rename(columns=rename_dict, inplace=True)
metadata_selected.head()

# subprocess create a sub-process to execute certain shell command
import subprocess

output_path_records = []
for file_id, row in metadata_selected.iterrows():
    tissue, dev_time, rep = row[['tissue', 'dev_time', 'replicate']]
    
    # assemble the input file path and make it absolute
    input_path = pathlib.Path(f'./data/small_fastq/{file_id}.fastq.gz').absolute()
    
    # assemble the output file path
    output_path = f'./data/fastq/{tissue}_{dev_time}_{rep}_{file_id}.fastq.gz'.replace(' ', '')
    
    # save the output file path
    output_path_records.append(pathlib.Path(output_path).name)
    
    # and create the soft link
    subprocess.run(['ln', '-s', input_path, output_path])
    
# add soft-link names to the metadata table
metadata_selected['file_name'] = output_path_records
metadata_selected.head()

' '.join(['ln', '-s', str(input_path), output_path])

# save the metadata table together with FASTQ files
metadata_selected.to_csv('data/fastq/fastq_metadata.csv')

## Output of this notebook

**In the ./data/fastq/ dir**

1. We have soft links of with meaningful names to the 25 original FASTQ files (named by ID from the database)
2. We have a clean metadata table recording all sample and database ID informations for these 25 files.

!ls -hl ./data/fastq/

metadata_selected

## Homework

Learn python pathlib package by yourself. (Hint: the most important class in this pacakge is the Path class, you don't need to learn everything about the whole package, try to search some simple tutorial on google)
