==> analysis/salmon_demo/step3.fastqTrimAndQC.ipynb <==
# FASTQ QC

## Aims of this notebook
- Trim the FASTQ by base quality, remove potential illumina adapter that may cause problem in mapping, and filter out short reads
- Update the metadata table using new file names

## Installation

### Install cutadapt, fastqc
`conda install -n genome_book -c bioconda cutadapt fastqc`

### Install trim_galore
see [trim galore github](https://github.com/FelixKrueger/TrimGalore) for instruction.
```{shell}
curl -fsSL https://github.com/FelixKrueger/TrimGalore/archive/0.6.5.tar.gz -o trim_galore.tar.gz
tar xvzf trim_galore.tar.gz
# Run Trim Galore
./TrimGalore-0.6.5/trim_galore
```

import pandas as pd
import pathlib
import subprocess

fastq_meta = pd.read_csv('./data/fastq/fastq_metadata.csv', index_col=0)
fastq_meta

output_dir = pathlib.Path('./data/fastq/trimmed/').absolute()
# another way to make dir
output_dir.mkdir(exist_ok=True)

fastq_dir = pathlib.Path('./data/fastq/').absolute()

## Run commands sequentially

# change this to your path to trim_galore installation
trim_galore_path = '/Users/hq/Documents/pkg/TrimGalore-0.6.5/trim_galore'

# iterate through each file in the fastq_meta
for file_name in fastq_meta['file_name']:
    file_path = fastq_dir / file_name
    # make the command using f-string
    command = f'{trim_galore_path} {file_path} --fastqc -o {output_dir}'
    # execute the command using subprocess
    subprocess.run(command, shell=True, check=True)
    print(command)
    print(file_name, 'trim finished.')

## [OPTIONAL] Run commands in parallel
- When you have multiple samples, it is important to learn how to execute commands in parallel, so you can utilize power of your multi-core laptop or server.

# Based on the number of cores your computer has
process = 4

# this is one way to run things in parallel in python
from concurrent.futures import ProcessPoolExecutor, as_completed
import subprocess

trim_galore_path = '/Users/hq/Documents/pkg/TrimGalore-0.6.5/trim_galore'
with ProcessPoolExecutor(process) as executor:
    #executor is a process pool, it will control the number of processes run in parallel
    
    futures = {}
    for file_name in fastq_meta['file_name']:
        file_path = fastq_dir / file_name
        command = f'{trim_galore_path} {file_path} --fastqc -o {output_dir}'
        
        # executor.submit a job (run a function with certain parameter, here the function is subprocess.run) 
        # and return a future obj. The future obj refer to this job, save it into futures first
        future = executor.submit(subprocess.run, command, shell=True, check=True)
        futures[future] = file_name
    
    # as_complete determine which future is finished
    for future in as_completed(futures):
        # get back the file name associated to this future
        file_name = futures[future]
        
        # this line is important, it check wheter the job finished without error
        # it will also got the job (subprocess.run) return, here that's not important
        _ = future.result()
        
        print(file_name, 'trim finished.')

## Make metadata for trimmed fastq

# find out all the trimmed fastq, make a dict
fastq_list = list(output_dir.glob('*trimmed.fq.gz'))
print(len(fastq_list))
fastq_list[:5]

# replace the fastq path with trimmed ones
fastq_series = pd.Series({i.name.split('_')[3]: str(i) 
                          for i in fastq_list})
fastq_meta['file_name'] = fastq_series

fastq_meta

fastq_meta.to_csv('data/fastq/trimmed/trimmed_fastq_metadata.csv')

!ls -hl data/fastq/trimmed/

## Output of this notebook

**In the ./data/fastq/trimmed dir**

1. We have 25 trimmed FASTQ files
2. We have a updated metadata table recording all sample and database ID informations for these 25 files.
