==> analysis/salmon_demo/step4.mapping.ipynb <==
# Salmon Mapping

## Aim of this notebook
1. Use salmon to quantify all the 25 trimmed FASTQ files for 16 samples
2. Generate a metadata table for salmon quant tables (16 sample)

## Prepare mapping commands


import pandas as pd
import pathlib

fastq_meta = pd.read_csv('./data/fastq/trimmed/trimmed_fastq_metadata.csv', index_col=0)
fastq_meta

## Prepare salmon command for each sample

# output dir
output_dir = pathlib.Path('data/quant/').absolute()
output_dir.mkdir(exist_ok=True)

# set all the directories
index_dir = pathlib.Path('data/salmon_index/').absolute()

# salmon accept threads parameter to allow parallel acceleration
# Change this number based on cores you have in your computer
# also, because salmon run in parallel internally, we just run salmon commands one by one
threads = 4

## Key step: merge FASTQ files from same sample together in one salmon command

There are 25 FASTQ files for 16 samples. Some samples (e.g. ENCSR160IIN Replicate 2) have multiple FASTQ file. When quantify reads using Salmon, you have to provide FASTQ files for the same sample together.

In the cell bellow, I use a key function of pandas called pd.DataFrame.groupby() to do this. 
See pandas documentation for more about groupby function
groupby: https://pandas.pydata.org/pandas-docs/stable/user_guide/groupby.html#splitting-an-object-into-groups

this is a very important and frquent used function

# make command for each RNA-seq sample based on the metadata
commands = {}
for (tissue, time, rep), sub_df in fastq_meta.groupby(['tissue', 'dev_time', 'replicate']):
    
    fastq_paths_str = ' '.join(sub_df['file_name'])
    output_name = output_dir / f'{tissue}_{time}_{rep}.quant'
    
    # assemble the final command
    command = f'salmon quant -i {index_dir} -l A -r {fastq_paths_str} --threads {threads} --validateMappings -o {output_name}'
    commands[f'{tissue}_{time}_{rep}'] = command

command

## Run salmon

import subprocess
for name, command in commands.items():
    # once command is finished, you may want to keep a physical record, so you know its finished for sure
    # you can also use this physical to prevent rerun the command, if the execution stopped in some place
    if pathlib.Path(output_dir / name).exists():
        print('EXISTS', name)
        continue
    
    subprocess.run(command, shell=True, check=True, 
                   stdout=subprocess.PIPE, stderr=subprocess.PIPE, encoding='utf8')
    
    print('FINISH', name)
    with open(output_dir / name, 'w') as f:
        f.write('Oh Yeah')
    

commands

## Clean up the flag

# optional, delete the flag files
# for name in commands.keys():
#     subprocess.run(f'rm {output_dir / name}', shell=True)

## Make a metadata for salmon output

# find out all the trimmed fastq, make a dict
salmon_quant_list = list(output_dir.glob('**/quant.sf'))
print(len(salmon_quant_list))
salmon_quant_list[:5]

pd.read_csv(salmon_quant_list[0], nrows=10, sep='\t', index_col=0)

records = []
for path in salmon_quant_list:
    # parse all sample information from the file name
    # In further analysis, we don't necessarily need database ID anymore, 
    # so here I chose to keep the metadata minimum
    sample_id = path.parent.name[:-6]
    tissue, time, rep = sample_id.split('_')
    records.append([sample_id, tissue, time, rep, str(path)])

# generate 
salmon_metadata = pd.DataFrame(records, 
                               columns=['sample_id', 'tissue', 'dev_time', 'replicate', 'salmon_count_path']
                              ).set_index('sample_id')
salmon_metadata.to_csv('data/quant/salmon_metadata.csv')

salmon_metadata

## Output of this notebook
In ./data/quant/ directory

1. We have subdirectories for each sample generated by salmon quant, within each the sample

# tree is a special command allows pretty printing of directory structure, 
# if you don't have this command, search how to install tree for linux
# or, just use the ls command
!tree ./data/quant/

salmon_metadata
